Add files via upload
The Student Management System is a console-based application that simplifies the management of student records. This project allows users to efficiently store, retrieve, update, and delete student information in a structured manner. Built in Python, the system replaces the traditional manual handling of student data with an intuitive menu-driven interface.
It is ideal for small-scale use cases such as educational institutions or personal projects requiring basic data storage and manipulation.

Features:

1- Add Students: Input multiple student records, including name, roll number, department, and course.

2- Display All Records:View the details of all stored students in a clear format.

3- Search Students: Retrieve specific student details using the roll number.

4- Update Records: Modify an existing student's information by searching for their roll number.

5- Delete Records: Option to delete either all student records or a specific record by roll number.

6- User-Friendly Menu: A text-based menu to guide users through the application functionalities.

Technologies Used:

- Programming Language: Python
- Concepts Utilized: Lists, Dictionaries, Functions, Control Structures (loops, conditionals)